refactor: split GameProvider
